posteriors as float32

+++ InnerEye/ML/pipelines/inference.py
@@ -407,9 +407,10 @@ class InferenceBatch(CTImagesMaskedBatch):
         # of shape but with an added class dimension: Num patches x Class x Z x Y x X
         predictions = np.concatenate(predictions, axis=0)
 
-        # create posterior output for each class with the shape: Class x Z x Y x x
+        # create posterior output for each class with the shape: Class x Z x Y x x. We use float32 as these
+        # arrays can be big.
         output_image_shape = self.pipeline.get_variable(InferencePipeline.Variables.OutputImageShape)
-        posteriors = np.zeros(shape=[model_config.number_of_classes] + list(output_image_shape))
+        posteriors = np.zeros(shape=[model_config.number_of_classes] + list(output_image_shape), dtype=np.float32)
         stride = self.pipeline.get_variable(InferencePipeline.Variables.Stride)
 
         for c in range(len(posteriors)):
